requirements.txt
mysqlclient
Flask
Flask-SQLAlchemy
flask_wtf
sqlalchemy-migrate
flask_security
bcrypt
python-dotenv
email_validator
-e git+https://github.com/LCBRU/lbrc_flask.git@main#egg=lbrc_flask
